+++ zfit/pdf.py
@@ -9,6 +9,8 @@ __all__ = [
     "CrystalBall",
     "DoubleCB",
     "GeneralizedCB",
+    "GaussExpTail",
+    "GeneralizedGaussExpTail",
     "Gauss",
     "Uniform",
     "TruncatedGauss",
@@ -17,6 +19,7 @@ __all__ = [
     "Poisson",
     "QGauss",
     "ChiSquared",
+    "StudentT",
     "Chebyshev",
     "Legendre",
     "Chebyshev2",
@@ -60,6 +63,7 @@ from .models.dist_tfp import (
     LogNormal,
     Poisson,
     QGauss,
+    StudentT,
     TruncatedGauss,
     Uniform,
     WrapDistribution,
@@ -76,7 +80,13 @@ from .models.kde import (
     KDE1DimISJ,
 )
 from .models.morphing import SplineMorphingPDF
-from .models.physics import CrystalBall, DoubleCB, GeneralizedCB
+from .models.physics import (
+    CrystalBall,
+    DoubleCB,
+    GaussExpTail,
+    GeneralizedCB,
+    GeneralizedGaussExpTail,
+)
 from .models.polynomials import (
     Chebyshev,
     Chebyshev2,